helper/cloud_vat_helper/requirements.txt
# VAT Helper - Google Sheets → BigQuery
# Dependências para sincronização de dados de VAT

google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
gspread>=5.0.0
pandas>=2.0.0
pytz>=2023.0
pyarrow>=14.0.0
db-dtypes>=1.0.0

